feat: prevent products & ventes to be modified after first order

+++ application/src/main/java/fr/sqq/achatgroupe/application/handler/command/DeleteProductHandler.java
@@ -1,7 +1,9 @@
 package fr.sqq.achatgroupe.application.handler.command;
 
 import fr.sqq.achatgroupe.application.command.DeleteProductCommand;
+import fr.sqq.achatgroupe.application.port.out.OrderRepository;
 import fr.sqq.achatgroupe.application.port.out.ProductRepository;
+import fr.sqq.achatgroupe.domain.exception.ProductModificationForbiddenException;
 import fr.sqq.achatgroupe.domain.exception.ProductNotFoundException;
 import fr.sqq.mediator.CommandHandler;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -11,9 +13,11 @@ import jakarta.transaction.Transactional;
 public class DeleteProductHandler implements CommandHandler<DeleteProductCommand, Void> {
 
     private final ProductRepository productRepository;
+    private final OrderRepository orderRepository;
 
-    public DeleteProductHandler(ProductRepository productRepository) {
+    public DeleteProductHandler(ProductRepository productRepository, OrderRepository orderRepository) {
         this.productRepository = productRepository;
+        this.orderRepository = orderRepository;
     }
 
     @Override
@@ -21,6 +25,9 @@ public class DeleteProductHandler implements CommandHandler<DeleteProductCommand
     public Void handle(DeleteProductCommand command) {
         productRepository.findByIdAndVenteId(command.id(), command.venteId())
                 .orElseThrow(() -> new ProductNotFoundException(command.id()));
+        if (orderRepository.existsNonCancelledByVenteId(command.venteId())) {
+            throw new ProductModificationForbiddenException();
+        }
         productRepository.deleteById(command.id());
         return null;
     }
